scm,git: start progress bar rendering at first task insertion
a "task" is a bare in rich progressbar widget.
If there is no task inserted, do not start rendering, otherwise,
a linefeed is inserted at "stop", even if nothing is rendered.
Leading to extra blank line in console output.

This is, for instance, the case is git repository is already up-to-date.