config: update config validation schema.
According to planed evolution for Cargo package support,
configuration slightly changes:

The followings nodes are renamed:
 - sentry --> kernel
 - libshield --> runtime (will cover runtime for supported languages)
 - app --> application

The following properties are renamed:
 - exelist --> provides (which may contain other compile product than an executable)
 - config_file --> config
 - deps --> depends

Scm and git configuration are no longer flattened.
each package has a scm.<method>, only scm.git is supported yet.

Application has a build.backend entry with meson/cargo value.
A generic options list can be passed w/ build.options.<opt>, they can be
either a boolean, a string or a integer value.

There will be a set of "built-in" options derived according to the backend.
Others, will be translate as-is and pass to configure step.
e.g.:

 toml:
   build.options.with_doc = true

 -->

 meson:
  -Dwith_doc=true