package,cargo: add custom install target
Add an internal for generic install target.
Use it for cargo as cargo can't add suffix to generated elf.
This is common to use elf extention for mcu firmware, in the other
hand, this is not needed for executable in a rich environment.

A lot of (more or less) recent build system allows to define target
suffix but cargo.